+++ src/document/parse/infobox/citation.js
