+++ src/document/parse/page/disambig.js
@@ -1,4 +1,4 @@
-const i18n = require('../../data/i18n');
+const i18n = require('../../../data/i18n');
 const parse_links = require('../section/sentence/links');
 const template_reg = new RegExp('\\{\\{ ?(' + i18n.disambigs.join('|') + ')(\\|[a-z =]*?)? ?\\}\\}', 'i');
