+++ src/document/parse/preProcess/kill_xml.js
