+++ src/document/parse/section/sentence/formatting.js
