+++ src/document/parse/section/sentence/sentence-parser.js
@@ -5,7 +5,7 @@
 // Ignore periods/questions/exclamations used in acronyms/abbreviations/numbers, etc.
 // @spencermountain 2015 MIT
 'use strict';
-const abbreviations = require('../../../data/abbreviations');
+const abbreviations = require('../../../../data/abbreviations');
 const abbrev_reg = new RegExp('(^| )(' + abbreviations.join('|') + ')[.!?] ?$', 'i');
 const acronym_reg = new RegExp('[ |.][A-Z].? +?$', 'i');
 const elipses_reg = new RegExp('\\.\\.\\.* +?$');
