+++ src/fetch/index.js
@@ -1,7 +1,7 @@
 //grab the content of any article, off the api
 const fetch = require('node-fetch');
 const site_map = require('../data/site_map');
-const Document = require('../Document');
+const Document = require('../document/Document');
 // const redirects = require('../parse/page/redirects');
 const isNumber = /^[0-9]*$/;
 
@@ -65,4 +65,6 @@ const getPage = function(title, lang, callback) {
     p.catch(reject);
   });
 };
+
+
 module.exports = getPage;
